Refactor Json formatters
