+++ app/uk/gov/hmrc/individualsincomeapi/services/v1/IncomeService.scala
@@ -22,7 +22,7 @@ import javax.inject.{Inject, Named, Singleton}
 import org.joda.time.{Interval, LocalDate}
 import uk.gov.hmrc.http.{HeaderCarrier, Upstream5xxResponse}
 import uk.gov.hmrc.individualsincomeapi.connector.{DesConnector, IndividualsMatchingApiConnector}
-import uk.gov.hmrc.individualsincomeapi.domain.JsonFormatters._
+import uk.gov.hmrc.individualsincomeapi.domain.v1.JsonFormatters._
 import uk.gov.hmrc.individualsincomeapi.domain.v1.SandboxIncomeData.findByMatchId
 import uk.gov.hmrc.individualsincomeapi.domain.des.DesEmployments
 import uk.gov.hmrc.individualsincomeapi.domain.MatchNotFoundException
