+++ test/unit/uk/gov/hmrc/individualsincomeapi/controllers/v1/IncomeControllerSpec.scala
@@ -31,7 +31,7 @@ import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.auth.core.retrieve.EmptyRetrieval
 import uk.gov.hmrc.domain.EmpRef
 import uk.gov.hmrc.individualsincomeapi.controllers.v1.{LiveIncomeController, SandboxIncomeController}
-import uk.gov.hmrc.individualsincomeapi.domain.JsonFormatters.paymentJsonFormat
+import uk.gov.hmrc.individualsincomeapi.domain.v1.Payment.paymentJsonFormat
 import uk.gov.hmrc.individualsincomeapi.domain.v1.SandboxIncomeData.sandboxMatchId
 import uk.gov.hmrc.individualsincomeapi.domain.MatchNotFoundException
 import uk.gov.hmrc.individualsincomeapi.domain.v1.Payment
